do not delete wallet state keys  starts with _

+++ client/src/pages/Wallet.jsx
@@ -49,6 +49,9 @@ const Wallet = () => {
   useEffect(() => {
     const keys = Object.keys(wallet)
     for (const k of keys) {
+      if (k.startsWith('_')) {
+        continue
+      }
       if (!apis.web3.isValidAddress(k)) {
         console.log(`Deleting stale wallet ${k}`)
         dispatch(walletActions.deleteWallet(k))
